pyproject.toml
[project]
name = "jarqyndos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nest-asyncio==1.6.0",
    "python-dotenv==1.0.1",
    "python-telegram-bot==21.10",
    "requests==2.32.3",
]
